Update patch set 1
Patch Set 1:

> It looks like the BIOS loader sets it for CDs and plenty of other media besides floppies: https://xref.landonf.org/source/xref/haiku/src/system/boot/platform/bios_ia32/shell.S#94
> 
> The PXE loader sets it unconditionally: https://xref.landonf.org/source/xref/haiku/src/system/boot/platform/pxe_ia32/devices.cpp#80
> 

They both set it in the case where they were loaded from outside a partition:
- The PXE loader is always doing things like that, because the partition is not loaded by the PXE ROM and is only found later (probably on a remote_disk_server)
- The bios_ia32 loader does it when started from the "floppy" entry point which will be only when it is outside of a BFS partition

When the flag is not set, it means that the bootloader knows which partition to boot from.
The bios_ia32 bootloader knows this thanks to makebootable, which allows it to know where it is located on disk.

Now for EFI, the bootloader is always outside the partition (it is in the EFI system partition). However, with the changes tqh is working on, it now has better self-awareness, that is, it knows which device it was booted from, and is able to locate BFS partitions on that same device. So, the idea is that it should pass on this information to the kernel
(in the form of partition offsets) and the kernel will be able to find the correct partition.

> The kernel uses two completely different sets of logic for partition detection based on whether or not it is set: https://xref.landonf.org/source/xref/haiku/src/system/kernel/fs/vfs_boot.cpp#264

Yes, when the flag is not set, the kernel trusts the bootloader. When the flag is set, the kernel attempts to guess a suitable partition. We know that the code to identify our CD image is broken for the 64bit version (the partition count check is incorrect, countchildren here will never be equal to 1: https://xref.landonf.org/source/xref/haiku/src/system/kernel/fs/vfs_boot.cpp#282). So it will fallback to this: https://xref.landonf.org/source/xref/haiku/src/system/kernel/fs/vfs_boot.cpp#301 that is, list all BFS and ISO1660 partitions (we should probably remove ISO1660 since we provide only anyboot CD images now).

Since this does not set foundForSure = true; , all partitions will be enumerated. Then, a sorting function is used to pick one. If the bootloader says we are booting from CD, the CD partitions will be used first, otherwise, we try to find something named "Haiku" or a name starting with "System" (both case insensitive), and if that also fails, finally we pick the first partition in ASCIIbetical order.

So, what this change does is disable all this fallback code in all cases when booting from EFI. Note that this fallback code was already disabled when booting from HDD or USB, since the EFI bootloader only enabled it when booting from read-only media. What the previous change you have linked did was always enable this fallback code and ignore any hints from the bootloader.

So I can conclude:
- Currently, booting using EFI from USB or HDD is working fine
- This change will have no impact in these cases
- Currently, booting using EFI from CD is unknown because it does not get a lot of testing (who is still using CDs and DVDs?)
- For the EFI loader, I see no reason for CDs to be different from USB or HDD, so I don't see why we could not use the same codepath
- For all non-EFI cases, nothing changes and nothing needs to be tested
- For all non-readonly cases (most likely anything other than CD will be read/write), nothing changes and nothing needs to be tested

Note that the effect of this change is moving CD booting from a codepath that we know for sure to be broken (at least because of the check on the number of partitions on the CD), to a codepath we know for sure is working in other very similar cases (EFI booting from HDD or USB).

Could it go wrong? Yes, but it seems unlikely, and we can fix it if it does. The only way to know is wide testing by many users with different configurations. And this is what nightly builds are for. If people are running that instead of beta releases, they have to accept that thet are on a very experimental branch and things break from time to time. Otherwise, we cannot ever merge anything. We can be unhappy about this, but then the answer is something like having more automated tests, or making more beta releases so people stop using nightlies in production. It is not blocking changes from getting into the master branch just because they might break things. If it breaks, we'll fix it, that's all.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>