Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Could it go wrong? Yes, but it seems unlikely, and we can fix it if it does.
> 
> My listing above was for the ideal set of things to test with. I do not think that is a requirement here; I think the only requirement is that HDD booting, USB booting, and CD booting are all tested (CD/DVD booting at least in a VM.)
> 
> If those have already been tested by tqh, then I have no objections.

Of course I've tested. It improves things but do not fix everything. We still don't do selection or select boot disk when we have several BFS partitions. It's just a small step in the right direction.

We don't have to pass in all partitions, I just wanted to make smaller changesets so we can improve and fix things quickly. One of the bigger problems I think was the partition check in kernel being wrong due to changed no of partitions, and the image flag which this fixes. By doing small fixes it is easy to detect, review and correct things, the big changes that laid the groundwork is done. Some regressions was from that change, some from adding the EFI partition on the anyboot iso.

In regards of process, my understanding is same as Pulkomandy, and I believe a VETO should be avoided unless there is a big issue springing from review.

When we voted for gerrit we agreed as long as people could still commit directly to the tree. I did so up until the big change I wanted a proper review on. I might have done one or two before that, and if I'd known how to set the WIP flag I'd done it for that change. Several mistakes were made.

I value your input and opinions a lot, but in this case I'd prefer if we just focus on the code changed.

Patch-set: 1
Reviewer: Gerrit User 1000021 <1000021@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>