Update patch set 4
Patch Set 4: Code-Review-1

As Adrien has noted before, even without the multiple million lines of generated code, DRM drivers are 200-400k LoC. I find it hard to believe we will be able to do with any less than that; and if we indeed cannot, I hesitate to say that should be in kernel space, and if it should not be in kernel space, does it or should it even be in-tree?

I think merging this with no real overall gameplan is premature. I also know X512 has talked about attempting to compile the OpenBSD drivers. We should discuss this at least on the mailing lists before going anywhere fast...

Patch-set: 4
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1