Update patch set 4
Patch Set 4:

> Yeah, it does conflict with https://review.haiku-os.org/c/haiku/+/430 from 2015 by Hamish. (it also adds all of the same drm headers) That whole series needs completely reworked and adds a *lot* more including an incomplete linux compatibility layer.
> 
> I think the approach in this change is a bit easier to digest since we can slowly merge the drm stuff on a per-card basis without impacting functionality *and* have multiple people working on it.

I don't understand this statement, the work Hamish did in 2015 took a very similar approach. In fact he had not touched the video card drivers at all in his patch series. He had focused on adding the code required to build the Linux drivers. I don't know if we want to go that way, or roll our own. My understanding of the 3D stack is that the kernel side driver is supposed to be as simple as possible, with the main part of the code being on Mesa side (similar to how our graphics drivers put most of the code in the accelerant). My understanding of the Linux code and later developments such as KMS (kernel modesetting) is that they have given up on this approach and put everything back in the kernel again. In any case, I think Hamish's work is generally useful if we want to try to use Linux drivers (as we do for FreeBSD drivers) and we should see what's reusable from there.

Maybe you are mixing things up with another attempt by Duggan and Vivek Roy who had taken a different approach with no plans for compatibility with existing video drivers.

Generally I tend to agree with Waddlesplash that I would try to put less things in kernel space. However, I don't know to what extent that's possible, and I think we are all a bit unfamiliar with the details of how DRM works and what exactly is required to be kernel-side.

So I think this is a good way to go about this: start with something similar to Linux, and get that working. Then, we will be able to identify the things that can be moved into userspace, and make something better than Linux. But we have to start from somewhere.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>