Update patch set 4
Patch Set 4:

> Generally I tend to agree with Waddlesplash that I would try to put less things in kernel space. However, I don't know to what extent that's possible, and I think we are all a bit unfamiliar with the details of how DRM works and what exactly is required to be kernel-side.

Actually I think it will be easier to do things in userspace here. We really should sit down and define more APIs for writing device drivers in userspace, and this sounds like the perfect time to do that. The only things we are really missing are interrupts handling and better PCI access systems.

But overall, the DRM drivers are so big and complex that working with them in userspace will be much easier than kernel space, especially with respect to debugging. FreeBSD's ports are not especially stable even in kernel space, and they have a team dedicated to managing them...

Patch-set: 4
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>