Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> > Generally I tend to agree with Waddlesplash that I would try to put less things in kernel space. However, I don't know to what extent that's possible, and I think we are all a bit unfamiliar with the details of how DRM works and what exactly is required to be kernel-side.
> 
> Actually I think it will be easier to do things in userspace here. We really should sit down and define more APIs for writing device drivers in userspace, and this sounds like the perfect time to do that. The only things we are really missing are interrupts handling and better PCI access systems.
> 
> But overall, the DRM drivers are so big and complex that working with them in userspace will be much easier than kernel space, especially with respect to debugging. FreeBSD's ports are not especially stable even in kernel space, and they have a team dedicated to managing them...

We have to put DRM card access in kernel space. Mesa is userspace and contains the majority of the rendering code for video cards. Linux has memory management calls within the drivers.

Essentially it is laid out like this:

* drm headers define command and control ioctl's used by kernel and userspace
* kernel drivers implement some card-specific ioctls for memory management (command ioctls)
  See https://cgit.freedesktop.org/mesa/drm/tree/include/drm/mga_drm.h#n233
  See https://cgit.freedesktop.org/mesa/drm/tree/include/drm/virtgpu_drm.h
* kernel drivers implement some generic drm ioctls for memory management (control)
  See https://cgit.freedesktop.org/mesa/drm/tree/include/drm/drm.h#n816
     The plan is to handle as many as possible via drm_common.c
* mesa finds the graphics device via /dev/graphics/* and checks for DRM and uses these memory management calls.

The linux model *is* memory management in the drivers, and rendering in userspace.

The big difference in my work and Hamish's is i'm focusing on getting the base of the API laid down so we can work at it in small chunks, Hamish was copying 300,000+ lines of DRM drivers into Haiku and trying to get them to build with Linux compatibility headers.

I think what Hamish was doing *does* need done at some point... but I think we should focus on the base (video card drivers, and presenting drm to user space) vs trying to cram as much Linux code as possible in all at once.

Patch-set: 5
Reviewer: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>