Update patch set 5
Patch Set 5:

> So then... the remaining argument against this patchset overall... Waddlesplash has been pushing for more of a "complete drm emulation" like we do for network cards.

Do you either think (1) that the 200-400k LoC (not including generated code) per GPU driver on Linux is superfluous, or (2) that we can replicate it? Because if the answers are "no", then we have to go the emulation route.

Not even the BSDs work from scratch on DRM drivers. It is madness to think that we can.

> The scope of change needed on that design would be massive (even it it was more ideal) and move us away from the BeAPI pretty quickly.

How?

BeOS never had hardware-accelerated graphics, at least until R5 (beyond the basics of "2D acceleration.") GPUs are also radically different than they were in the 90s. The accelerants API is showing its age and needs to be replaced.

Patch-set: 5
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>