Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > So then... the remaining argument against this patchset overall... Waddlesplash has been pushing for more of a "complete drm emulation" like we do for network cards.
> 
> Do you either think (1) that the 200-400k LoC (not including generated code) per GPU driver on Linux is superfluous, or (2) that we can replicate it? Because if the answers are "no", then we have to go the emulation route.
> 
> Not even the BSDs work from scratch on DRM drivers. It is madness to think that we can.

I love Haiku because we dare to be "mad". It started when some people gathered and decided to rewrite an entire OS from scratch. It should continue this way.

The advantages of this:

1) We learn a lot about how drivers actually work and what they do
2) We get a chance to learn from the mistakes of Linux instead of blindly copypasting them

If you think porting the Linux drivers is easier, fine, you can try it. But please don't waste your time and that of other contributors by arguing that they should not be trying different approaches. If this doesn't work, fine, nothing is lost, we can drop it and try another way.

Also, Linux is written in C, from my experience this will be about a 5x increase in count of lines of code when compared to C++. And yes, we might not need all the features they have in their drivers, too, so maybe something much simpler is possible. Let's find out!

Patch-set: 5
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>