Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> I rather have this fast than have to remake both USB and this.
> 
> Where are we. Does it work? I have more Bluetooth preference to do before I can test this.

You would need yet another copy of the HID files for Bluetooth so this change does not make you closer to anything for Bluetooth (it is useful only for I2C touchpads).

The minimal change that's needed here is putting these common files into a shared library, and linking the 3 drivers (USB, I2C and Bluetooth HID) to that library.

Later on we can see about making HID a separate driver/module instead, if that's useful.

Patch-set: 7