Update patch set 1
Patch Set 1:

> This seems like more of a workaround vs fixing the actual kernel bug?

Yes, my understanding (please confirm) is:

- Something disable interrupts (we're not sure what)
- Normally, interrupts should only be disabled while running kernel code, but in this case, we somehow escape this
- This part of the code happens to check the interrupt state, so we crash here.

I wonder if interrupts are actually disabled, however. I expect that both nodejs and dosbox will be doing strange things to the stackframes, and possibly corrupting them in some way. Could that be the case? So the stackframe ends up not being in the format we expect because the app is doing strange things with the stack, then a page fault occurs, and then it crashes.

Since we have only triggered this problem with these two specific programs, it seems likely that something from userland is triggering the problems.

In the case of Dosbox, setting it to not use "dynamic" mode (where instructions from the emulated system are run on the host) may avoid the problem, confirming that dosbox somehow leaks something from the emulated machine when running in this mode (https://www.dosbox.com/wiki/Configuration:CPU)

Patch-set: 1