Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Now it work better.
> 
> * Edit field growing/shrinking works.

This is the main thing I was focusing on so good.

> * Horizontal scrolling is not working correctly if text view right part is outside window: cursor will be always at text view left side.

Can you give me an example of where this is happening? The cursor works fine for me with horizontal scrolling in StyledEdit even if the right part of the text is outside of the window. Could this have something to do with different testing procedures?

> * Text still jumps by 1-2 pixels when pressing F2. Not critical. Font is DejaVu Sans 12.

I see what you mean here: F2 performs Edit name (I use Cmd+E and didn't know F2 worked here) and I've seen the text rect off a pixel or 2 in various places when the text view switches to being editable. Calculation must be off somewhere in BTextWidget::StartEdit. It's not perfect but it is livable.

> * Text fields in Beezer settings window are still scrolled vertically so content is almost invisible (https://dev.haiku-os.org/attachment/ticket/16642/Beezer BTextView bug.png).

This is a bug in Beezer, the bottom of the text rect is set to 0. We could work around this in BTextView though I suppose.

Patch-set: 4
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>