Update patch set 2
Patch Set 2:

Just as a reminder, we'll also need to point the FTP network_settings add-on to the tnftpd executable over here, update AboutSystem not to mention ftpd, and update the Haiku Userguide not to mention ftpd too. (Just wanted to create a quick list! :D)

Also, I see that there is another ftpd instance here in the tree:
https://github.com/haiku/haiku/tree/master/src/bin/network/ftpd

Patch-set: 2
CC: Gerrit User 1000052 <1000052@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>