Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > The alignment of 8-byte fields to 8-byte address may not be cross-platform?
> 
> Right, this would fix on 32-bit. But isn't this code only 64-bit on Haiku?

For zircon, it only supports x86_64 and arm64, this is not the problem with Pad field.
The compiler will handle this with alignment. This would be the issue with arm, and 
the legacy 32bits platform.

Patch-set: 1
Reviewer: Gerrit User 1000301 <1000301@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>