Update patch set 1
Patch Set 1: Code-Review-2

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > The alignment of 8-byte fields to 8-byte address may not be cross-platform?
> > 
> > Right, this would fix on 32-bit. But isn't this code only 64-bit on Haiku?

Very nice find.

This padding can't be in the header as it's alignment per platform. If you look at EFI spec p. 165 in 2.8 Final:
  typedef struct {
    UINT32 Type ;
    EFI_PHYSICAL_ADDRESS PhysicalStart ;
    EFI_VIRTUAL_ADDRESS VirtualStart ;
    UINT64 NumberOfPages ;
    UINT64 Attribute ;
  } EFI_MEMORY_DESCRIPTOR;

It needs to follow the padding rules for different archs. It is also specified in the spec.
Therefore we need to do this in a different way.

Patch-set: 1
Reviewer: Gerrit User 1000021 <1000021@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2