Update patch set 1
Patch Set 1:

> This padding can't be in the header as it's alignment per platform. If you look at EFI spec p. 165 in 2.8 Final

Indeed.  It's weird though that GNUEFI has that Pad *and* it fixes our 32-bit booting issues (nice find btw!)

Maybe Pad is somehow fixing this requirement on the same page?

" PhysicalStart   Physical address of the first byte in the memory region. PhysicalStart must be aligned on a 4KiB boundary, and must not be above 0xfffffffffffff000. Type EFI_PHYSICAL_ADDRESS is defined in the AllocatePages() function description."


Or maybe on page 23...

"Any UEFI memory descriptor that requests a virtual mapping via the EFI_MEMORY_DESCRIPTOR having the EFI_MEMORY_RUNTIME bit set must be aligned on a 4KiB boundary and must be a multiple of 4KiB in size"

Maybe a __attribute__((aligned(4096))) on the EFI_MEMORY_DESCRIPTOR struct or something?

Patch-set: 1
CC: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>