Update patch set 1
Patch Set 1:

an alternative strategy might be trying to detect a "detached serial_puts" and changing the serial target to printf... but that would be a behavior change on all architectures / cases where we may not want it.

A common source of issue when debugging bootloader issues is lack of a serial port on modern x86 hardware... this debug define method might help there as well.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>