Update patch set 1
Patch Set 1:

> > 
> > And if you have the menu working you can go to debug options -> show bootloader log to see the log, does that work?
> 
> Not if you want to see all the log output up to loading the kernel :-)
> 
> > Also, this function calls serial_puts, shouldn't that already get things to the serial port
> 
> Not in practice.  I think the stdio "screen" is directed at the serial port, thus the serial output gets directed to nowhere (you can't have multiple consumers of the serial port in our loader?)

This worked perfectly for me for the sparc/openfirmware version.

Early logs get to the console. Logs during the menu init and use of the menu get in the menu bootloader log.

After exiting the menu, you get back to logging to console.

So if EFI isn't doing that properly, is it possible to fix it? Instead of overlapping menu and serial logs in the same place? The problem is the same on non-serial platforms, by the way, the console and menu always share the same output, screen or serial port, it doesn't matter.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>