Update patch set 5
Patch Set 5:

> > Patch Set 5:
> > 
> > Can you rebase this? I am not sure about changes with pixel offset of Tracker name edit field.
> 
> I can't rebase until the redeclaration of C function conflicts are resolved in BeOSBuildCompatibility.h.

There's not much to fix here. If you have a recent checkout (which I guess you have), it will build on Haiku when your system is hrev54993 or later. If you want to build a post hrev55013 source tree on a pre-hrev54993 image, then you need to manually revert the declarations in BeOSBuildCompatibility.h.

Patch-set: 5
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>