Update patch set 7
Patch Set 7: Code-Review+1

Thanks for working on this :)

The changes look ok to me. Should we merge this and use it in nightlies for a while to see if we find more problems, or do you want to give it more testing on your own first?

Patch-set: 7
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1