Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why not use stack guard pages? Compiler has an option to probe stack pages when allocating a large memory block on stack.

We already use stack guard pages. They allow to grow the stack as needed.

The stack protector is not designed for the same use. What it does is make sure no one attempts to erase data in the stack (for example by overflowing an array), intentionally or by accident.

The way this works is:
- Tell the compiler that stack protection is enabled
- The compiler will call our functions to add a marker at the top of the stack (before pushing the context of the next function), and check that marker when exiting functions
- If some code overwrites the marker, we know that a stack problem occured.

Note that on x86, the stack grows downwards. So here is a possible layout with function 1 calling function 2:

So the memory layout is like this:

^ Towards address FFFFFFFF

<- Start of stack
<- Function 1 frame pointer
Function 1 variables go here
Return address goes here
<- Function 2 frame pointer
Stack protector for function 2 goes here
Function 2 local variables go here

Free space for the stack is here

Guard page is here, to detect when we need more stack memory (it triggers a page fault when the stacxk grows past the space already allocated to it)

v Towards address 00000000

Now if some array is allocated in function 2 and the code writes past its end, it will erase the return address and then the variables of function 1 (because the end of the array goes towards larger addresses, not towards the top of the stack which is at smaller addresses). It will not get to the guard page, which is before the start of the array.

Why a random value? This just makes it harder for malicious code (or malicious input to existing code) to know in advance what the stack protection value is. It makes attacks that manage to somehow erase parts of the stack more complicated to execute (now they first need to find out the value so they can write the same one back, or find a way to rewrite parts of the stack while jumping over the stack protection markers).

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>