Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> To make sure i understood you correctly: do you mean i should push the whole gcc source tree into gerrit?

No, because there is no need to review it all.

Currently the code used to build Haiku is the one from our buildtools repository, because it is easy to build on any OS. This repository is self-contained with all dependencies, limiting problems when building on various other platforms.

The recipe at haikuports should be similar to that repository because we want the compiler it generates to behave the same and be compatible with it.

So we should keep that in mind. It does not matter if we do the recipe first and the buildtools repo later or the other way around, but we should think about how to make sure the resulting compilers (build from buildtools repo and built from the recipe) are as similar as possible.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>