Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > @extrowerk you might as well copy the script as build_cross_tools_gcc11, avoiding more conflicts..
> 
> Why can't we just fix the setting so it works for both CLang building crosstools and GCC11. Duplicating just adds even more work for adminstrators. Just ask diver or someone else with CLang to verify that it works. I just don't think that you should regress already working combinations.

Everyone setting up a haiku cross-compiler on any OS will use this script, so it needs to be very general. On FreeBSD the cross-compiler will be built by CLang, on Debian it will be an older GCC and on some 'cool' Linux distros it will be a newer. If you break things a lot of interested developers will have a hard time getting started and give up, so we should be able to do both.

Patch-set: 1
Reviewer: Gerrit User 1000021 <1000021@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>