Update patch set 2
Patch Set 2:

I like the small font, too, on my 1920x1200 display. Of course, that depends on your preferences, and abilities.

I would prefer to only enable it when the resolution is larger than that (best in both directions to make sure it's not a wide screen, ie. 3840x1080 or something like that), and also have a kernel setting to enforce either way.

Patch-set: 2