Update patch set 1
Patch Set 1:

> Aren't we waiting on a new icu for webkit?

That change request could be merged. Only "problem" remaining in it is that Python 3 depends on readline7 instead of the current version of readline on 64bit systems. So there is one extra useless package to add to the build.

Discussion is ongoing with haikuports team to fix that.

The icu package we are "waiting" on has been ready for years now. But no one ever added it to the haiku packages, and now I'm trying to do that. But because no one updated anything else either, I end up doing a lot more work than I expected: newer version of ffmpeg, freetype with harfbuzz support, migration from python2 to python3, ...

All of this creating extra work and extra delays.

We need to do releases more often so these things don't rot as much and need so much work to get them back working...

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>