Update patch set 1
Patch Set 1:

This may workaround a regression from BeOS R5 which had no SetInsets() yet still managed to position the text rect with the right insets in StyledEdit.

We don’t update the insets based on the initial text rect passed into the constructor because the initial text rect is often not set correctly. You could argue that this is a bug and that we should update the insets in the constructor but this would break many existing apps so I am not doing that.

We could probably call SetTextRect() here instead of SetInsets after text view construction and that would update the insets. However, SetInsets() works reliably to give us the border around the text in the text view on all sides while SetTextRect() sets the text rect (and thus the insets) differently based on word-wrap setting while SetInsets() always does the right thing regardless of the word-wrap setting.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>