Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Wouldn't it make more sense to merge this with change 53, since that one hasn't been merged yet?
> 
> I prefer a commit of my own :) It is meant to fix shortcomings of the change 53, but not only.

To me it's fine to replace an existing change with a new one made in your own name (just reuse the same change-id). The "owner" in Gerrit will not change, but that is not really used once the patch is merged.

Well, it doesn't really matter. Should we abandon change 53 then?

Patch-set: 3
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>