Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> 1. keep this arm macro and leave it for rest platforms.

I vote for this for now (and it is unrelated to this commit). I think that in a long term kernel args should be converted into key-value pairs like in Zircon kernel to allow better flexibility and extensibility[1]. Current hardcoded structs are bad solution because it is difficult to extend without breaking boot loader <-> kernel protocol.

1. https://github.com/vsrinivas/fuchsia/blob/4f31fb23cea1cd0531808bd96c3adffc010bcf9f/src/firmware/gigaboot/src/zircon.c#L275

Patch-set: 4