Update patch set 5
Patch Set 5:

I'd like to see the people that has issues with Beta 2 try nightlies before this approach. The partition mismatch in VFS check caused a lot of them, and some might have been due to poor method used by EFI-loader as well. The fixes was post Beta2.

This just add the same code as what platform_get_boot_partition should already do. I'd prefer to error with a lot of logging about passed in boot partition when we fail to find one as this is more of a band-aid than finding the root-cause why VFS didn't match with the loader.

Patch-set: 5
CC: Gerrit User 1000021 <1000021@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>