Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > And without any version, this is reckless.
> 
> Here's all of the things that I thought about before I made this one-line change:

Thanks, it's easier to understand what you're aiming at now.

Personally I'm fine with Debian as well, but it definitely should be a fixed version. This way, the container is in a "known to work" setup. When a new Debian (or Ubuntu) version comes, we have to test it before we update the container. Otherwise, we never know which version we're getting, and the container is useless since it doesn't guarantee a fixed, known-good environment (what's the point, then?).

So, yes, unifying on a single OS sounds like a good idea, using slim/minimal versions is a good idea, and Debian vs Ubuntu is a personal choice so no need to discuss that further, pick whichever you like.

Patch-set: 1