Update patch set 2
Patch Set 2:

As Adrien explained, these scripts are supposed to provide an environment which just works. Because there is already no version constraints on the installed packages, it means it defaults to what the distribution provides (we can't pin every package version). The image has to be correctly defined. IMO Ubuntu 17.04 was a precise version enough. Instead of buster, i would use "10.9".

Patch-set: 2