Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> As Adrien explained, these scripts are supposed to provide an environment which just works. Because there is already no version constraints on the installed packages, it means it defaults to what the distribution provides (we can't pin every package version). The image has to be correctly defined. IMO Ubuntu 17.04 was a precise version enough. Instead of buster, i would use "10.9".

Using "buster" is similar to "17.04" in Ubuntu terms. The difference between point releases of Debian is only bugs and security fixes. Using an older point release does not change this, since as soon as you use apt to install something, it will use the latest version from the Buster repository anyway.

Debian is stable enough that this is normally not a problem.

Patch-set: 2