Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (3 comments)

* --head=50 reduces the required amount bandwidth, to the benefit of the user. Is having a full git history necessary for an environment that's just meant to build stuff? (I think I'm leaning towards "Yes", now that I'm thinking about it, because it could be used for development too, especially since that's the direction I'm planning to take with the Docker images.)
* As far as removing the echo is concerned, echo wouldn't show output with the commands supplied in the Makefile. I tried further debugging the `prep.sh` script by adding a lot of echo messages, and that wasn't very helpful, so I just removed all of them, including the one you're referring to, even though I removed that one accidentally. Should I add all of the echoes back?
* Debugging reasons, it's a leftover and I should probably move it back, even if there shouldn't be a difference, practically speaking. I put anything concerning a subdirectory of a directory to the bottom, in order to better understand why Docker wasn't able to find the `jam` folder and compile it. Needless to say, `buildtools` does seem to get cloned, I think, but it can't find jam?

I'd be more than glad to iron out those details before this gets merged, but my general lack of direction as far as to what/why is any of this happening is pretty much preventing me from further investigating the issues. I am not even sure if the images are generally broken or if I'm the only one dealing with all of this. I'm really just trying to get things to work, since these things don't seem to be working, and that's why I'm taking the course of direction that I did. Does any of this even actually work for you all, or are you able to reproduce any of the things I brought up?

Patch-set: 3
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>