Update patch set 1
Patch Set 1:

I have restored the original version of the script on the builder.

This change does not work because the version comparison ignores the "revision" part. So for example it picked python 3.7.10-1 over python 3.7.10-4.

The existing code works just fine on the docker builders. I am not sure what problem you ran into, but it's now fixed.

Abandoning the change.

Patch-set: 1