Update patch set 2
Patch Set 2:

> This problem can be solved by centering CPU name and frequency together with logo.

It is not that easy because the code doesn't know which part of the logo is used. All logos are HVIF icons and they have to be square. Current logos are "packed" to the top of the square with empty space at the bottom, and text overlaps the empty space.

But I'll find a way to make it fit :)

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>